fix(db): folder position should be bigint, not int4 (Postgres)
Folder `position` is written as a millisecond epoch timestamp (~1.75e12),
which overflows a 32-bit Postgres `integer` and would fail folder creation
on Postgres deployments with "value out of range for type integer". It now
matches created_at/updated_at, which are already bigint.

Since the 0012 migration has never been applied to a real Postgres database,
fold the type change into a regenerated 0012 (via drizzle-kit) rather than
adding a separate 0013 alter, keeping the Postgres and SQLite migration
histories aligned at 0012. SQLite's `position` is already a 64-bit INTEGER,
so its schema and migrations are unchanged.

+++ packages/database/migrations/postgres/0012_user_additions.sql
@@ -2,7 +2,7 @@ CREATE TABLE "user_folders" (
 	"id" text NOT NULL,
 	"user_id" text NOT NULL,
 	"name" text NOT NULL,
-	"position" integer DEFAULT 0 NOT NULL,
+	"position" bigint DEFAULT 0 NOT NULL,
 	"created_at" bigint DEFAULT (extract(epoch from now()) * 1000)::bigint,
 	"updated_at" bigint DEFAULT (extract(epoch from now()) * 1000)::bigint,
 	CONSTRAINT "user_folders_id_user_id_pk" PRIMARY KEY("id","user_id")
